eatgo-customer-api/src/main/java/com/soten/eatgo/reservation/service/ReservationService.java
package com.soten.eatgo.reservation.service;

import com.soten.eatgo.reservation.domain.Reservation;
import com.soten.eatgo.reservation.domain.ReservationRepository;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Service
@Transactional
public class ReservationService {

    private ReservationRepository reservationRepository;

    public ReservationService(ReservationRepository reservationRepository) {
        this.reservationRepository = reservationRepository;
    }

    public Reservation addReservation(Long restaurantId, Long userId, String name, String date, String time, Integer partySize) {
        Reservation reservation = Reservation.builder()
                .restaurantId(restaurantId)
                .userId(userId)
                .name(name)
                .date(date)
                .time(time)
                .partySize(partySize)
                .build();
        return reservationRepository.save(reservation);
    }
}

